requirements.txt
pip install -r requirements.txt

flask
flask-cors
Pillow
